add anthropic.ts
